Added a details page and a link back to root on header
Filling in details of the Details.js, we now have the details for a specific pet when clicked.

The Pet.js passes in the props necessary and added a link.

App.js added a link back to the home page in the <header>

Results.js passes in an id for use in Pet.js/Details.js

R

Signed-off-by: Stephen Weiss <stephencweiss@gmail.com>